Add task 52

+++ task50/Program.cs
@@ -18,7 +18,7 @@ int m = Convert.ToInt32(Console.ReadLine());
 Console.Write("Введите  n: ");
 int n = Convert.ToInt32(Console.ReadLine());
 
-double [,] matrix = new double [m,n];
+int [,] matrix = new int [m,n];
 for(int i = 0; i < m; i++)
 {
     for(int j = 0; j < n; j++)
